perf: isolate messages panel logic exports
